account-decoder: Add feature flags to reduce dependency tree
Add optional feature flags allowing selective inclusion of account decoders. Users can now depend on specific features (config, stake, vote, etc.) instead of pulling the entire dependency tree.

- Move shared types to always-available core module
- Gate each parse_* module behind its respective feature
- Gate parse_account_data and encode_ui_account behind full feature
- Re-export types from parse_account_data for backward compatibility
- Default to full feature for backward compatibility

Fixes #10102
