+++ account-decoder/src/parse_address_lookup_table.rs
@@ -1,5 +1,5 @@
 use {
-    crate::parse_account_data::{ParsableAccount, ParseAccountError},
+    crate::core::{ParsableAccount, ParseAccountError},
     serde::{Deserialize, Serialize},
     solana_address_lookup_table_interface::state::AddressLookupTable,
     solana_instruction::error::InstructionError,
