+++ account-decoder/src/parse_nonce.rs
@@ -1,5 +1,5 @@
 use {
-    crate::{parse_account_data::ParseAccountError, UiFeeCalculator},
+    crate::core::{ParseAccountError, UiFeeCalculator},
     serde::{Deserialize, Serialize},
     solana_instruction::error::InstructionError,
     solana_nonce::{state::State, versions::Versions},
